backend/requirements.txt
flask
flask-cors
python-dotenv
pandas
openpyxl
plotly
google-generativeai